+++ src/src.pro
@@ -24,14 +24,16 @@ SOURCES += \
     sakura_parsing/sakura_parser_interface.cpp \
     sakura_parsing.cpp \
     sakura_items.cpp \
-    value_item_map.cpp
+    value_item_map.cpp \
+    sakura_garden.cpp
 
 HEADERS += \
     sakura_parsing/sakura_parser_interface.h \
     ../include/libKitsunemimiSakuraParser/sakura_parsing.h \
     ../include/libKitsunemimiSakuraParser/sakura_items.h \
     ../include/libKitsunemimiSakuraParser/value_items.h \
-    ../include/libKitsunemimiSakuraParser/value_item_map.h
+    ../include/libKitsunemimiSakuraParser/value_item_map.h \
+    ../include/libKitsunemimiSakuraParser/sakura_garden.h
 
 FLEXSOURCES = grammar/sakura_lexer.l
 BISONSOURCES = grammar/sakura_parser.y
